# openapi-spec.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SStats.net - Football API",
    "description": "API футбольных матчей, доматчевые и live коэффициенты, статистика, профиты, Glicko 2\r\n\r\nВесь функционал API бесплатен, с ограничением по количеству запросов. \r\nБез api ключа ограничение 300 в минуту запросов на всех пользователей, или 30 запросов в минуту на 1 IP адрес.\r\n\r\n[Введение в API](https://sstats.net/api)",
    "contact": {
      "name": "Поддержка",
      "url": "https://vk.com/sstats_net"
    },
    "version": "0.9.13.0"
  },
  "servers": [
    {
      "url": "https://api.sstats.net",
      "description": "SStats.net API"
    }
  ],
  "paths": {
...
    "/Teams/list": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Список команд",
        "description": "Список команд с возможностью фильтрации и пагинации",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "description": "Название команды",
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "name": "Country",
            "in": "query",
            "schema": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 0,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список команд с информацией о пагинации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfApiSaTeam[]"
                }
              }
            }
          }
        }
      }
    },
    "/Teams/{id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Получает детальную информацию о команде по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор команды",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfApiSaTeamFull"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfObject"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfApiSaTeam[]": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ApiSaTeam"
            }
          },
          "requestQuery": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "TotalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiResponseOfApiSaTeamFull": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaTeamFull"
              }
            ]
          },
          "requestQuery": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "TotalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfObject": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "data": {},
          "requestQuery": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "TotalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiSaCountry": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Уникальный код страны"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ApiSaLeague": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaCountry"
              }
            ]
          },
          "flashScoreId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiSaPlayer": {
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ApiSaSeason": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "league": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaLeague"
              }
            ]
          }
        }
      },
...
      "ApiSaTeam": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "flashId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Id команды с сайта Flashscore"
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaCountry"
              }
            ]
          }
        }
      },
      "ApiSaTeamFull": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "flashId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Id команды с сайта Flashscore"
          },
          "logoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaCountry"
              }
            ]
          },
          "code": {
            "type": [
              "null",
              "string"
            ],
            "description": "Сокращенная аббревиатура команды"
          },
          "founded": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "Год основания команды",
            "format": "int32"
          },
          "seasons": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ApiSaSeason"
            },
            "description": "Сезоны команды"
          },
          "venue": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Стадион",
                "$ref": "#/components/schemas/ApiSaVenue"
              }
            ]
          },
          "coach": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiSaPlayer"
              }
            ]
          },
          "players": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ApiSaPlayer"
            }
          }
        }
      },
      "ApiSaVenue": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "address": {
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "type": [
              "null",
              "string"
            ]
          },
          "capacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiAuthScheme": {
        "type": "apiKey",
        "description": "API key в query string, например: ?apikey=<my_api_key>",
        "name": "apikey",
        "in": "query"
      }
    }
  },
...
}